+++ common/src/main/java/io/github/jamalam360/reaping/mixin/VillagerEntityMixin.java
@@ -32,8 +32,8 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @Mixin(VillagerEntity.class)
 public abstract class VillagerEntityMixin extends MerchantEntity implements CustomReapableEntityDuck {
-    private boolean reapingmod$remainSmall = false;
-    private int reapingmod$remainingSmallTicks = 0;
+    private boolean reaping$remainSmall = false;
+    private int reaping$remainingSmallTicks = 0;
 
     public VillagerEntityMixin(EntityType<? extends MerchantEntity> entityType, World world) {
         super(entityType, world);
@@ -43,29 +43,29 @@ public abstract class VillagerEntityMixin extends MerchantEntity implements Cust
             method = "writeCustomDataToNbt",
             at = @At("TAIL")
     )
-    public void reapingmod$serializeRemainSmall(NbtCompound nbt, CallbackInfo ci) {
-        nbt.putBoolean("reapingmod$remainSmall", this.reapingmod$remainSmall);
-        nbt.putInt("reapingmod$remainSmallTicks", this.reapingmod$remainingSmallTicks);
+    public void reaping$serializeRemainSmall(NbtCompound nbt, CallbackInfo ci) {
+        nbt.putBoolean("reaping$remainSmall", this.reaping$remainSmall);
+        nbt.putInt("reaping$remainSmallTicks", this.reaping$remainingSmallTicks);
     }
 
     @Inject(
             method = "readCustomDataFromNbt",
             at = @At("TAIL")
     )
-    public void reapingmod$deserializeRemainSmall(NbtCompound nbt, CallbackInfo ci) {
-        this.reapingmod$remainSmall = nbt.getBoolean("reapingmod$remainSmall");
-        this.reapingmod$remainingSmallTicks = nbt.getInt("reapingmod$remainSmallTicks");
+    public void reaping$deserializeRemainSmall(NbtCompound nbt, CallbackInfo ci) {
+        this.reaping$remainSmall = nbt.getBoolean("reaping$remainSmall");
+        this.reaping$remainingSmallTicks = nbt.getInt("reaping$remainSmallTicks");
     }
 
     @Inject(
             method = "tick",
             at = @At("TAIL")
     )
-    public void reapingmod$tick(CallbackInfo ci) {
-        if (this.reapingmod$remainingSmallTicks != 0 && this.reapingmod$remainSmall) {
-            this.reapingmod$remainingSmallTicks--;
-        } else if (this.reapingmod$remainingSmallTicks <= 0 && this.reapingmod$remainSmall) {
-            this.reapingmod$remainSmall = false;
+    public void reaping$tick(CallbackInfo ci) {
+        if (this.reaping$remainingSmallTicks != 0 && this.reaping$remainSmall) {
+            this.reaping$remainingSmallTicks--;
+        } else if (this.reaping$remainingSmallTicks <= 0 && this.reaping$remainSmall) {
+            this.reaping$remainSmall = false;
             ReapingExpectPlatform.setScale(this, 1f);
         }
     }
@@ -75,19 +75,19 @@ public abstract class VillagerEntityMixin extends MerchantEntity implements Cust
             at = @At("HEAD"),
             cancellable = true
     )
-    public void reapingmod$reapVillager(PlayerEntity player, Hand hand, CallbackInfoReturnable<ActionResult> cir) {
+    public void reaping$reapVillager(PlayerEntity player, Hand hand, CallbackInfoReturnable<ActionResult> cir) {
         if (ReapingHelper.tryReap(player, this, player.getStackInHand(hand)) == ActionResult.SUCCESS) {
             cir.setReturnValue(ActionResult.SUCCESS);
         }
     }
 
     @Override
-    public ActionResult reapingmod$onReaped(@Nullable PlayerEntity user, ItemStack toolStack) {
+    public ActionResult reaping$onReaped(@Nullable PlayerEntity user, ItemStack toolStack) {
         int lootingLvl = EnchantmentHelper.getLevel(Enchantments.LOOTING, toolStack);
 
-        if (!this.reapingmod$remainSmall) {
-            this.reapingmod$remainSmall = true;
-            this.reapingmod$remainingSmallTicks = ReapingMod.RANDOM.nextInt(50 * 20, 120 * 20);
+        if (!this.reaping$remainSmall) {
+            this.reaping$remainSmall = true;
+            this.reaping$remainingSmallTicks = ReapingMod.RANDOM.nextInt(50 * 20, 120 * 20);
 
             this.dropStack(new ItemStack(ReapingItems.HUMAN_MEAT.get(), lootingLvl == 0 ? 1 : this.world.random.nextInt(lootingLvl) + 1));
             ReapingExpectPlatform.setScale(this, 0.45f);