+++ common/src/main/java/io/github/jamalam360/reaping/registry/ReapingTrades.java
@@ -11,42 +11,42 @@ import net.minecraft.village.VillagerProfession;
  */
 public class ReapingTrades {
     public static void registerTrades() {
-        TradeRegistry.registerVillagerTrade(
+        ReapingItems.IRON_REAPER.listen((item) -> TradeRegistry.registerVillagerTrade(
                 VillagerProfession.BUTCHER,
                 2,
                 new SimpleTrade(
                         new ItemStack(Items.EMERALD, 3),
                         ItemStack.EMPTY,
-                        ReapingItems.IRON_REAPER.get().getDefaultStack(),
+                        item.getDefaultStack(),
                         5,
                         10,
                         1
                 )
-        );
+        ));
 
-        TradeRegistry.registerVillagerTrade(
+        ReapingItems.DIAMOND_REAPER.listen((item) -> TradeRegistry.registerVillagerTrade(
                 VillagerProfession.BUTCHER,
                 5,
                 new SimpleTrade(
                         new ItemStack(Items.EMERALD, 13),
                         ItemStack.EMPTY,
-                        ReapingItems.DIAMOND_REAPER.get().getDefaultStack(),
+                        item.getDefaultStack(),
                         3,
                         30,
                         1
                 )
-        );
+        ));
 
-        TradeRegistry.registerTradeForWanderingTrader(
+        ReapingItems.HUMAN_MEAT.listen((item) -> TradeRegistry.registerTradeForWanderingTrader(
                 true,
                 new SimpleTrade(
                         new ItemStack(Items.EMERALD, 7),
                         ItemStack.EMPTY,
-                        ReapingItems.HUMAN_MEAT.get().getDefaultStack(),
+                        item.getDefaultStack(),
                         6,
                         20,
                         1
                 )
-        );
+        ));
     }
 }