+++ forge/src/main/java/io/github/jamalam360/reaping/forge/ReapingForge.java
@@ -1,19 +1,17 @@
 package io.github.jamalam360.reaping.forge;
 
+import com.mojang.serialization.Codec;
 import dev.architectury.platform.forge.EventBuses;
 import io.github.jamalam360.reaping.ReapingMod;
 import io.github.jamalam360.reaping.config.ReapingConfig;
-import io.github.jamalam360.reaping.registry.ReapingEntities;
 import me.shedaniel.autoconfig.AutoConfig;
-import net.minecraft.entity.SpawnGroup;
-import net.minecraft.world.biome.SpawnSettings;
-import net.minecraft.world.gen.feature.StructureFeatures;
-import net.minecraftforge.client.ConfigGuiHandler;
-import net.minecraftforge.event.world.StructureSpawnListGatherEvent;
-import net.minecraftforge.eventbus.api.SubscribeEvent;
+import net.minecraftforge.client.ConfigScreenHandler;
+import net.minecraftforge.common.world.StructureModifier;
 import net.minecraftforge.fml.ModLoadingContext;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
+import net.minecraftforge.registries.DeferredRegister;
+import net.minecraftforge.registries.ForgeRegistries;
 
 @Mod(ReapingMod.MOD_ID)
 public class ReapingForge {
@@ -22,15 +20,12 @@ public class ReapingForge {
         ReapingMod.init();
 
         ModLoadingContext.get().registerExtensionPoint(
-                ConfigGuiHandler.ConfigGuiFactory.class,
-                () -> new ConfigGuiHandler.ConfigGuiFactory((minecraftClient, screen) -> AutoConfig.getConfigScreen(ReapingConfig.class, screen).get())
+                ConfigScreenHandler.ConfigScreenFactory.class,
+                () -> new ConfigScreenHandler.ConfigScreenFactory((client, screen) -> AutoConfig.getConfigScreen(ReapingConfig.class, screen).get())
         );
-    }
 
-    @SubscribeEvent
-    public void onStructureSpawnListGatherEvent(StructureSpawnListGatherEvent event) {
-        if (event.getStructure() == StructureFeatures.PILLAGER_OUTPOST.get()) {
-            event.addEntitySpawn(SpawnGroup.MONSTER, new SpawnSettings.SpawnEntry(ReapingEntities.REAPING_PILLAGER.get(), 1, 1, 1));
-        }
+        DeferredRegister<Codec<? extends StructureModifier>> serializers = DeferredRegister.create(ForgeRegistries.Keys.STRUCTURE_MODIFIER_SERIALIZERS, ReapingMod.MOD_ID);
+        serializers.register(FMLJavaModLoadingContext.get().getModEventBus());
+        serializers.register("spawn_modifier", ReapingStructureModifier::makeCodec);
     }
 }